Split apply logic into edit_op_apply.rs to fix audit GodFile/HighItemCount
Move resolve_anchor(), apply_edit_ops_to_content(), apply_edit_ops(),
ApplyReport, ApplyError, and their 24 tests into a dedicated module.

edit_op.rs: 520 lines (types + conversions + 9 tests)
edit_op_apply.rs: 1073 lines (apply logic + 24 tests)
